[feat] add signers info in signatures endpoint (#1882)

+++ server/src/pdf/__types.ts
@@ -5263,7 +5263,7 @@ export type UpdatePetitionInput = {
 };
 
 export type UpdatePetitionVariableInput = {
-  defaultValue: Scalars["Int"]["input"];
+  defaultValue: Scalars["Float"]["input"];
 };
 
 export type UpdateProfileFieldValueInput = {
